Utilizo módulos ESM y pulo detalles
