+++ index.html
@@ -17,8 +17,8 @@
         </div>
 
         <span class="toolbar_span-menu">
-          <img src="./assets/img/menu.png"
-        /></span>
+          <img src="./assets/img/menu.png" />
+        </span>
 
         <div class="toolbar_nav">
           <a href="">
@@ -31,8 +31,8 @@
           </a>
           <a href="">
             <img src="./assets/img/aboutMe.png" alt="" />
-            <p>About me</p></a
-          >
+            <p>About me</p>
+          </a>
         </div>
 
         <aside class="aside">aside lista</aside>
@@ -53,22 +53,13 @@
       </section>
 
       <div class="highlights" id="DOMcardsProjects">
-        <!-- <section class="card">
-            <h3>Proyecto</h3>
-            <div ><img  class="card_img"src="/assets/img/avatarDavina.png" alt=""></div>
-            <p>
-              Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.
-            </p>
-            <p><b>Skills:</b> html, css, js </p>
-            <a href="#" class="card__btn">Source </a>
-            <a href="#" class="card__btn optional">Demo </a>
-          </section> -->
+        <!-- Los cards se generan en Javascript -->
       </div>
     </main>
     <footer>
       <img src="/assets/img/logoDavina.pink.png" type="module" alt="" />
     </footer>
 
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
