+++ js/data_projects.js
@@ -1,39 +1,74 @@
-export const dataProjects = [
+export default [
   {
     id: 1,
-    title: " Toca el Dombo ",
-    src: "/assets/img/projects/domboo.png",
-    description: " Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
+    title: "Toca el Dombo",
+    src: "./assets/img/projects/domboo.png",
+    description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
     skills: "CSS, HTML, JS DOM",
-    source: "https://github.com/Davina-dev/Toca-el-Dombo",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
     demo: "#"
   },
   {
     id: 2,
-    title: " Eloquent JS Solutions",
-    src: "/assets/img/projects/eloquentJs.png",
+    title: "Eloquent JS Solutions",
+    src: "./assets/img/projects/eloquentJs.png",
     description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
     skills: "CSS, HTML, JS DOM",
-    source: "https://github.com/Davina-dev/Toca-el-Dombo",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
     demo: "#"
   },
   {
     id: 3,
-    title: "  CRUD React",
-    src: "/assets/img/projects/crudReact.png",
+    title: "CRUD React",
+    src: "./assets/img/projects/crudReact.png",
     description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
     skills: "CSS, HTML, React Hooks, Node",
-    source: "https://github.com/Davina-dev/Toca-el-Dombo",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
     demo: "#"
   },
   {
     id: 4,
-    title: " PokeApi ",
-    src: "/assets/img/projects/pokeapi.png",
+    title: "PokeApi",
+    src: "./assets/img/projects/pokeapi.png",
     description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
     skills: "CSS, HTML, JS DOM",
-    source: "https://github.com/Davina-dev/Toca-el-Dombo",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
+    demo: "#"
+  },
+  {
+    id: 5,
+    title: "Calculator",
+    src: "/assets/img/projects/calculator.png",
+    description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
+    skills: "CSS, HTML, React Hooks, Node",
+    source: "https:github.com/Davina-dev/calculator",
+    demo: "#"
+  },
+  {
+    id: 6,
+    title: "Food for Home",
+    src: "./assets/img/projects/foodForHome.png",
+    description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
+    skills: "CSS, HTML, JS DOM",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
+    demo: "#"
+  },
+  {
+    id: 7,
+    title: "CRUD React",
+    src: "./assets/img/projects/crudReact.png",
+    description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
+    skills: "CSS, HTML, React Hooks, Node",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
+    demo: "#"
+  },
+  {
+    id: 8,
+    title: "PokeApi",
+    src: "./assets/img/projects/pokeapi.png",
+    description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Neque labore iure iusto perferendis, rerum mollitia voluptatibus eum molestias delectus officiis, atque quaerat quisquam ratione doloribus. Eos vero officia omnis est.",
+    skills: "CSS, HTML, JS DOM",
+    source: "https:github.com/Davina-dev/Toca-el-Dombo",
     demo: "#"
   }
-];
-
+]
