+++ js/load_cards.js
@@ -1,6 +1,10 @@
+import { printCards } from './print_cards.js';
+import dataProjects from './data_projects.js';
+
 export function loadCards() {
+  let DOMcardsProjects = document.getElementById("DOMcardsProjects");
   DOMcardsProjects.innerHTML = "";
-  data_projects.forEach((data, i) => printCards(data, i));
+  dataProjects.forEach((data, i) => printCards(data, i));
 }
 
-export const saludo = "hola , soy un modulo";
+export const saludo = "Hola, soy Davina y lo peto!";
